+++ public/pages/workflow_detail/agentic_search/components/no_deployed_models_callout.tsx
@@ -7,17 +7,18 @@ import React from 'react';
 import { EuiCallOut, EuiLink } from '@elastic/eui';
 import { AGENTIC_SEARCH_DOCS_LINK } from '../../../../../common';
 
+interface NoDeployedModelsCalloutProps {
+  title?: string;
+}
+
 /**
  * Basic callout component with links to documentation.
  */
-export function NoDeployedModelsCallout({}) {
+export function NoDeployedModelsCallout({
+  title = 'No deployed models found',
+}: NoDeployedModelsCalloutProps) {
   return (
-    <EuiCallOut
-      size="s"
-      color="warning"
-      iconType={'alert'}
-      title="No deployed models found"
-    >
+    <EuiCallOut size="s" color="warning" iconType={'alert'} title={title}>
       Deploy models compatible with agentic search. For more information, see
       the{' '}
       <EuiLink target="_blank" href={AGENTIC_SEARCH_DOCS_LINK}>