+++ public/pages/workflow_detail/agentic_search/configure_flow/agent_configuration.tsx
@@ -547,6 +547,7 @@ export function AgentConfiguration(props: AgentConfigurationProps) {
                                 <AgentAdvancedSettings
                                   agentForm={props.agentForm}
                                   setAgentForm={props.setAgentForm}
+                                  isNewAgent={props.newAndUnsaved}
                                 />
                               </EuiFlexItem>
                             </>