+++ public/route_service.ts
@@ -26,6 +26,7 @@ import {
   BULK_NODE_API_PATH,
   BASE_NODE_API_PATH,
   SEARCH_CONNECTORS_NODE_API_PATH,
+  SEARCH_MEMORY_CONTAINERS_NODE_API_PATH,
   GET_MAPPINGS_NODE_API_PATH,
   SEARCH_PIPELINE_NODE_API_PATH,
   INGEST_PIPELINE_NODE_API_PATH,
@@ -107,6 +108,7 @@ export interface RouteService {
   }) => Promise<any>;
   searchModels: (body: {}, dataSourceId?: string) => Promise<any>;
   searchConnectors: (body: {}, dataSourceId?: string) => Promise<any>;
+  searchMemoryContainers: (body: {}, dataSourceId?: string) => Promise<any>;
   registerAgent: (body: {}, dataSourceId?: string) => Promise<any>;
   updateAgent: (
     agentId: string,
@@ -437,6 +439,19 @@ export function configureRoutes(core: CoreStart): RouteService {
         const url = dataSourceId
           ? `${BASE_NODE_API_PATH}/${dataSourceId}/connector/search`
           : SEARCH_CONNECTORS_NODE_API_PATH;
+        const response = await core.http.post<{ respString: string }>(url, {
+          body: JSON.stringify(body),
+        });
+        return response;
+      } catch (e: any) {
+        throw e;
+      }
+    },
+    searchMemoryContainers: async (body: {}, dataSourceId?: string) => {
+      try {
+        const url = dataSourceId
+          ? `${BASE_NODE_API_PATH}/${dataSourceId}/memory_container/search`
+          : SEARCH_MEMORY_CONTAINERS_NODE_API_PATH;
         const response = await core.http.post<{ respString: string }>(url, {
           body: JSON.stringify(body),
         });